Extract converters from __init__ file
